fixed bug where the io plugin would seg fault while opening feather files

+++ io/src/feather.hpp
@@ -62,7 +62,7 @@ namespace io
         struct data_t {
             unsigned int fid;
             unsigned short type;
-            unsigned short length;
+            uint32_t length;
         };
 
         bool open(std::string filename);
